# gateway/src/main/java/com/cozisoft/starter/business/core/model/UpdateAccountStatusRequest.java
package com.cozisoft.starter.business.core.model;

import lombok.AllArgsConstructor;
import lombok.Builder;
import lombok.Data;
import lombok.NoArgsConstructor;
import lombok.experimental.Accessors;

@Data
@Builder
@NoArgsConstructor
@AllArgsConstructor
@Accessors(chain = true)
public class UpdateAccountStatusRequest {
    private AccountStatusUpdateAction action;
}